Add make_error_code helpers for errors that require static_casts to shorten the code

+++ ouzel/core/windows/EngineWin.cpp
@@ -188,8 +188,8 @@ namespace ouzel
         if (MultiByteToWideChar(CP_UTF8, 0, url.c_str(), -1, buffer.data(), buferSize) == 0)
             throw std::system_error(GetLastError(), std::system_category(), "Failed to convert UTF-8 to wide char");
 
-        intptr_t result = reinterpret_cast<intptr_t>(ShellExecuteW(nullptr, L"open", buffer.data(), nullptr, nullptr, SW_SHOWNORMAL));
+        int result = reinterpret_cast<int>(ShellExecuteW(nullptr, L"open", buffer.data(), nullptr, nullptr, SW_SHOWNORMAL));
         if (result <= 32)
-            throw std::system_error(static_cast<int>(result), shellExecuteErrorCategory, "Failed to execute open");
+            throw std::system_error(result, shellExecuteErrorCategory, "Failed to execute open");
     }
 }
